requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
matplotlib
numpy
jinja2
aiofiles
